+++ VintageMagicMod/VintageMagicModModSystem.cs
@@ -3,6 +3,7 @@ using Vintagestory.API.Common;
 using Vintagestory.API.Config;
 using Vintagestory.API.Server;
 using Vintagestory.GameContent;
+using VintageMagicMod.WitchOven;
 
 namespace VintageMagicMod
 {
